pyproject.toml
[project]
name = "the-learning-monkey"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "google-adk>=1.1.1",
    "ipykernel>=6.29.5",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.63",
    "langchain-google-genai>=2.1.5",
    "pydantic>=2.11.5",
]
